Added update for last name
I had to use a second form in the profile page so the input for first
name and last name could be differentiated on the server. This will
probably need to be refactored at some point. DB test for update will
definitely need to be redone to accurately test functionality.