Redo on profile.html page
My original idea for the user's profile page was to display data in a
table. I changed this to display all of the user's information in an
uneditable form. I also originally had 2 forms on the profile page, one
for first name and another for last name. Each could be edited
individually which meant there was a separate form validator class for
each. Going forward you can edit all your profile information and the
full form will be validated and unique keys validated DB side.